Tweaks to the CSXL Slack Button
