+++ frontend/src/app/navigation/navigation.component.ts
@@ -11,7 +11,7 @@ import { ActivatedRoute, Router, RouterEvent } from '@angular/router';
 import { Profile, ProfileService } from '../profile/profile.service';
 import { PermissionService } from '../permission.service';
 import { NagivationAdminGearService } from './navigation-admin-gear.service';
-import { SlackInviteBox } from '../shared/slack-invite-box/slack-invite-box.widget';
+import { SlackInviteBox } from './widgets/slack-invite-box/slack-invite-box.widget';
 
 @Component({
   selector: 'app-navigation',
@@ -77,8 +77,7 @@ export class NavigationComponent implements OnInit, OnDestroy {
 
   onSlackInviteClick(): void {
     const dialogRef = this.slackDialog.open(SlackInviteBox, {
-      width: '1000px',
-      autoFocus: false
+      autoFocus: 'dialog'
     });
     dialogRef.afterClosed().subscribe();
   }
