+++ frontend/src/app/navigation/widgets/slack-invite-box/slack-invite-box.widget.css
