--- frontend/src/app/shared/slack-invite-box/slack-invite-box.widget.html
@@ -1,35 +1,0 @@
-<mat-card class="slack-dialog" #dialogContainer>
-  <mat-card-content class="content-container">
-    <h1>Join us on Slack!</h1>
-    <p>
-      Interested in becoming a part of the Undergraduate Computer Science
-      Experience Lab community?
-    </p>
-    <p>
-      Join the CSXL Slack to get updates about the XL, learn about new
-      opportunities for CS students, or meet your fellow CS majors!
-    </p>
-    <br />
-    <h3>
-      By accepting the below invite and joining the CSXL Slack, you are agreeing
-      to the following:
-    </h3>
-    <p>
-      I will abide by the rules and standards detailed in the CSXL Community
-      Standards Agreement.
-    </p>
-    <p>
-      In addition, upon joining the CSXL Slack, I will complete my Slack profile
-      by adding my real name and an appropriate profile picture of myself.
-    </p>
-    <div class="slack-button-container">
-      <button
-        mat-stroked-button
-        class="slack-button"
-        onclick="window.open('https://join.slack.com/t/csxl/shared_invite/zt-1mvfaiqme-fXXw9cKjaXOfhXVBfgXVCA','_blank')"
-        type="button">
-        Join the Slack!
-      </button>
-    </div>
-  </mat-card-content>
-</mat-card>
