--- frontend/src/app/shared/slack-invite-box/slack-invite-box.widget.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'slack-invite-box',
-  templateUrl: './slack-invite-box.widget.html',
-  styleUrls: ['./slack-invite-box.widget.css']
-})
-export class SlackInviteBox {
-  constructor() {}
-}
